Almost working panel driver (DANGER BEWARE)
This revision drives the pixels in a way which
works, but only if you never stop updating and
shifting the addressed pixel. If it pauses for more than
a couple of seconds your panel will start to smoke and burn
this is because it is not pulsing the 18v line and relying
on you having drive it high manually.